Add utils/overwrite_writer.h, needs a little bit more work
